Update CONTRIB.md
Reason of district was to assist in understanding if each district still requires full lockdown mode or not . High vaccinated region can be exclusively be opened to restart our economy while the opposite should remain in enforced mode. 

This way we can reduce frontliner's effort in enforcing in all areas.